Add files via upload
Title: Components of a Circuit Board on a Blue Surface
Photographer: Karolina Grabowska
Website: https://www.pexels.com/photo/components-of-a-circuit-board-on-a-blue-surface-7286032/
Date: 02 November 2023
Description: A close-up view of electronic components on a circuit board with a blue surface.
License: Pexels License
Usage Rights: Pexels Rights - Free to use, even for commercial purposes. No attribution required.