Add files via upload
Title: The Curious Incident of the Dog in the Night-Time
Creator: Mark Haddon
Subject: Fiction, Mystery, Coming-of-age
Date: 2003
Type: Book
Format: Paperback
Identifier: ISBN 038572010X
Source: Doubleday
Language: English
Description: The Curious Incident of the Dog in the Night-Time is a novel by British author Mark Haddon. It was first published in 2003 and has won numerous awards, including the Whitbread Book Award and the Carnegie Medal. The novel tells the story of Christopher Boone, a 15-year-old boy with autism, who investigates the murder of his neighbor's dog. Besides the murder he discovered more than intended for example is parents getting divorced. This is one of the first books I have ever read and always go back to has it has a sense of murder mystery that I enjoy as well as personal experiances that I can relate to.  
Coverage: Global
Rights: Copyright © 2003 Mark Haddon. All rights reserved.