Add files via upload
This is a video posted by YouTuber called The Bobby Show and I used this tutorial to not only learn how to use garageband a music based application for work related purposes as this is how I recored, edit and combined the interview. arageBand is a digital audio workstation (DAW) developed by Apple Inc. It allows users to create, record, edit, and mix music. GarageBand comes pre-installed on all macOS and iOS devices, making it a popular choice for both beginners and professional musicians. Its metadata elements are Title: The title of the song or project.
Creator: The name of the artist or musician who created the song or project.
Subject: The genre or style of music.
Description: A brief description of the song or project.
Publisher: Apple Inc.
Date: The date on which the song or project was created or last modified.
Type: The type of digital resource, such as "audio recording" or "musical composition."